move diffHashSig to frontend to remove d3 dependency from ambient-utils
