+++ src/ambient-utils/package.json
@@ -2,10 +2,10 @@
     "name": "@crocswap-libs/ambient-utils",
     "version": "0.1.0",
     "private": false,
+    "license": "MIT",
     "dependencies": {
         "@covalenthq/client-sdk": "^0.2.6",
         "@crocswap-libs/sdk": "^0.2.79",
-        "d3": "^7.4.4",
         "ethers": "^5.6.5",
         "numbro": "^2.3.6",
         "wagmi": "^0.12.12"
