+++ src/pages/Chart/FeeRate/FeeRateChart.tsx
@@ -6,10 +6,7 @@ import '../Chart.css';
 import { scaleData, setCanvasResolution } from '../ChartUtils/chartUtils';
 import { CandleDataIF } from '../../../ambient-utils/types';
 import { createIndicatorLine } from '../ChartUtils/indicatorLineSeries';
-import {
-    diffHashSig,
-    diffHashSigScaleData,
-} from '../../../ambient-utils/dataLayer';
+import { diffHashSig, diffHashSigScaleData } from '../../../utils/diffHashSig';
 
 interface FreeRateData {
     feeData: Array<CandleDataIF>;
