+++ src/types/ScaleDataIF.ts
